Support removal of attributes from type dict after deletion
